feat: double check to delete a guest

+++ src/app/room/[id]/page.tsx
@@ -425,22 +425,46 @@ export default function EditRoom({ params }: { params: { id: string } }) {
                       {/* NamesPool */}
                       <div className="flex flex-wrap gap-2 text-white">
                         {initialRoom?.guests?.map((guest) => (
-                          <button
-                            onClick={(e) => {
-                              e.preventDefault();
-                              handleDeleteGuest(guest.id, guest.name);
-                            }}
-                            key={guest.id}
-                          >
-                            <span
-                              className={` ${
-                                guest.isReady ? `bg-primary` : `bg-gray`
-                              } flex items-center justify-center gap-1 rounded-2xl px-3 py-1 text-lg`}
-                            >
-                              {guest.name}
-                              <X size={12} />
-                            </span>
-                          </button>
+                          <AlertDialog.Root key={guest.id}>
+                            <AlertDialog.Trigger asChild>
+                              <button onClick={() => null} key={guest.id}>
+                                <span
+                                  className={` ${
+                                    guest.isReady ? `bg-primary` : `bg-gray`
+                                  } flex items-center justify-center gap-1 rounded-2xl px-3 py-1 text-lg`}
+                                >
+                                  {guest.name}
+                                  <X size={12} />
+                                </span>
+                              </button>
+                            </AlertDialog.Trigger>
+                            <AlertDialog.Portal>
+                              <AlertDialog.Overlay className="bg-blackA6 data-[state=open]:animate-overlayShow fixed inset-0" />
+                              <AlertDialog.Content className="data-[state=open]:animate-contentShow fixed left-1/2 top-1/2 max-h-[85vh] w-[90vw] max-w-[500px] -translate-x-1/2 -translate-y-1/2 rounded-md bg-white p-[25px] shadow-[hsl(206_22%_7%_/_35%)_0px_10px_38px_-10px,_hsl(206_22%_7%_/_20%)_0px_10px_20px_-15px] focus:outline-none">
+                                <AlertDialog.Title className="m-0 mb-4 text-[17px] font-medium text-black">
+                                  {`Are you absolutely sure your desire is delete the guest ${guest.name}?`}
+                                </AlertDialog.Title>
+                                <div className="flex justify-end gap-[25px]">
+                                  <AlertDialog.Cancel asChild>
+                                    <button className="inline-flex h-[35px] items-center justify-center rounded-xl border border-black bg-white px-4 py-6 font-medium leading-none text-black outline-none">
+                                      Cancel
+                                    </button>
+                                  </AlertDialog.Cancel>
+                                  <AlertDialog.Action asChild>
+                                    <button
+                                      onClick={(e) => {
+                                        e.preventDefault();
+                                        handleDeleteGuest(guest.id, guest.name);
+                                      }}
+                                      className="inline-flex h-[35px] items-center justify-center rounded-xl border border-black bg-black px-4 py-6 font-medium leading-none text-white outline-none"
+                                    >
+                                      Yes, delete guest
+                                    </button>
+                                  </AlertDialog.Action>
+                                </div>
+                              </AlertDialog.Content>
+                            </AlertDialog.Portal>
+                          </AlertDialog.Root>
                         ))}
                       </div>
                     </div>
